requirements.txt
langchain==1.2.9
langchain-core==1.2.9
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
huggingface-hub>=0.24.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
streamlit>=1.32.0
python-dotenv>=1.0.1
pypdf>=4.0